+++ CHANGELOG.md
@@ -57,3 +57,9 @@
 ## 0.0.8
 
 * Refactor from method <code>bool PrinterStatus.isImpressoraOK()</code> to getter <code>bool PrinterStatus.isImpressoraOK</code>.
+
+## 0.0.9
+
+* Add the method <code>Future<PrinterStatus> imprimirCodigoDeBarraImg(String mensagem, int width, int height, BarcodeTypes barcodeType)</code>.
+* Remove the method <code>Future<PrinterStatus> fimImpressao()</code> and call it after every printer command.
+* Updated README.md and example.